Deleted CyclicBuffer and DynamicArray, created SnakesManager
DynamicArray was very hard to use and prevent memory leakage, so i decided to replace it with SnakesManager created only to manage snakes in GamePlayEngine. SnakesManager still has crytical bugs, they will be fixed in nearest commits. CyclicBuffer also was hard to use and when needed, there will be replacement for it too.